+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- public api for response objects
 ### Changed
 ### Deprecated
 ### Removed
 ### Fixed
 ### Security
+
+## [0.2.1] - 2021-04-11
+### Added
+- public api for response objects
 
 ## [0.2.0] - 2021-04-05
 ### Added